requirements.txt
fastapi
uvicorn
python-multipart
motor
python-jose[cryptography]
passlib[bcrypt]
python-dotenv